simplified the react app
