--- src/example/README.md
@@ -1,7 +1,0 @@
-## React service container
-
-This Example is from 
-
-https://github.com/traviskaufman/react-service-container
-
-That is the closest thing I found to what I wanted.
